ansible/requirements.txt
ansible==2.9.6
skale.py==3.10dev7
vultr==1.0.1
